+++ src/lib/app-fns/explorer/index.ts
@@ -12,6 +12,7 @@ export const getExplorerTxUrl = (chainName: string) => {
       pathSuffix = "txs";
       break;
     case "terra2":
+    case "terra2testnet":
       pathSuffix = "tx";
       break;
     default:
@@ -28,7 +29,23 @@ export const getExplorerUserAddressUrl = (chainName: string) => {
       pathSuffix = "account";
       break;
     case "terra2":
+    case "terra2testnet":
       pathSuffix = "address";
+      break;
+    default:
+      break;
+  }
+  return `${explorerMap[chainName]}/${pathSuffix}`;
+};
+
+export const getExplorerBlockUrl = (chainName: string) => {
+  let pathSuffix = "";
+  switch (chainName) {
+    case "osmosis":
+    case "osmosistestnet":
+    case "terra2":
+    case "terra2testnet":
+      pathSuffix = "blocks";
       break;
     default:
       break;