+++ src/lib/components/ExplorerLink.tsx
@@ -4,6 +4,7 @@ import { useWallet } from "@cosmos-kit/react";
 import Link from "next/link";
 
 import {
+  getExplorerBlockUrl,
   getExplorerTxUrl,
   getExplorerUserAddressUrl,
 } from "lib/app-fns/explorer";
@@ -15,7 +16,8 @@ export type LinkType =
   | "tx_hash"
   | "user_address"
   | "contract_address"
-  | "code_id";
+  | "code_id"
+  | "block";
 
 interface ExplorerLinkProps extends BoxProps {
   value: string;
@@ -45,6 +47,9 @@ const getNavigationUrl = (
       break;
     case "code_id":
       url = "/code";
+      break;
+    case "block":
+      url = getExplorerBlockUrl(currentChainName);
       break;
     default:
       break;